Extract session lifecycle logic into helpers
- Extracted complex logic from `src/hooks/session-lifecycle.ts` to `src/hooks/session-lifecycle-helpers.ts`.
- Created helper functions for stale session handling, hierarchy visualization, warning compilation, and anchor loading.
- Cleaned up `createSessionLifecycleHook` to focus on orchestration.
- Verified with integration tests and type check.
